+++ src/Commands/CommandRegistration.cs
@@ -1,6 +1,7 @@
 ﻿using System.ComponentModel.Composition;
 using Microsoft.VisualStudio.Editor;
 using Microsoft.VisualStudio.Text;
+using Microsoft.VisualStudio.Text.Classification;
 using Microsoft.VisualStudio.Text.Editor;
 using Microsoft.VisualStudio.TextManager.Interop;
 using Microsoft.VisualStudio.Utilities;
@@ -18,6 +19,9 @@ namespace MarkdownEditor
         [Import]
         ITextDocumentFactoryService TextDocumentFactoryService { get; set; }
 
+        [Import]
+        IClassifierAggregatorService ClassifierAggregatorService { get; set; }
+
         public void VsTextViewCreated(IVsTextView textViewAdapter)
         {
             var textView = EditorAdaptersFactoryService.GetWpfTextView(textViewAdapter);
@@ -26,6 +30,8 @@ namespace MarkdownEditor
             if (TextDocumentFactoryService.TryGetTextDocument(textView.TextBuffer, out document))
             {
                 textView.Properties.GetOrCreateSingletonProperty(() => new PasteImage(textViewAdapter, textView, document.FilePath));
+                textView.Properties.GetOrCreateSingletonProperty(() => new CommentCommandTarget(textViewAdapter, textView, ClassifierAggregatorService));
+                textView.Properties.GetOrCreateSingletonProperty(() => new UncommentCommandTarget(textViewAdapter, textView, ClassifierAggregatorService));
             }
         }
     }